Migrated regular distribution configs parsing to composable parser.
